~Optionally~ use <pre> (or any element) with specific id instead of <noscript>
Blank page is shown if javascript is enabled but md-page script fails to load
for some reason (for example if the linked script is unavailable).
This happens because html isnt generated when md-page.js fails and
browsers won't show the fallback <noscript> content if js is enabled.

<noscript> content is also seen garbled in browsers as
formatting including newlines and multiple spaces is ignored in html.

By using <pre> instead, we can guarantee display of md content and also
that formatting is preserved. An md-page specific element id can be used
to indicate exactly which element to parse.
For backwards compatibility <noscript> tag is also allowed.
The first match found of either an element with id="md-page-markdown-text"
or <noscript> will be used.

Co-Authored-By: harsgak <harsgak@users.noreply.github.com>